fix(script) Directory is the current working one.
Explicitely use the current working directory the script directory.

This behavior was already present before, but a bug fix (see
https://github.com/atoum/atoum/pull/770) has revealed this other
bug.

The script directory must not rely on the script name (which is defined
in the `atoum` binary by the `atoum\scripts\runner` constant). The
script name points to the binary name most of the time (it is the case
right now), but the script directory is the user current working
directory, they are not related.